WIP: Initial draft of addons based on Helm
Kubespray's mission is to deploy a production Kubernetes. Unfortunately,
not everyone has the same understanding on what is a production
Kubernetes. For some, a production Kubernetes is a bunch of "post-kubeadm" VMs. For others, a production Kubernetes has a lot of addons.

It has been discussed that the current support for addons in Kubespray
adds a huge maintenance burden, and reusing existing Helm Charts (via
Helm 3) is a desired direction -- assuming Kubespray should still tackle
addons.

This patch is initial draft for Helm-based addons support in Kubespray.
It is meant to stear discussions on the future of Kubespray addons.

My intention is to demonstrate a solution that:
- allows addons to be easily installed, configured,
upgraded;
- follows the Principle of Least Astonishment (POLA) --
a Kubespray user should feel that they would have probably ended up with
a similar solution, had they started from scratch;
- reduces maintenance burden on the Kubespray project;
- may attract contributors outside of Kubespray.

The high-level approach taken here is as follows. Kubespray provides an
umbrella addons Chart which includes sub-Charts of popular addons. Nginx-ingress is given as a first example of such an addon. By default, all addons are disabled. If enabled via the `helm_addons_enabled` Ansible variable, then the Chart is installed under the `kube-addons-system` namespace. The Release's values can be configured via the `helm_addons_values` Ansible variable.

The setup leaves the Chart and the produced values file on the first master node, to put the user in the driver's seat. Specifically, the user can SSH into the first master and do `helm upgrade` themselves until they obtain the desired not configuration, before updating the Ansible variables.

To test this patch, add the following variables:

```
helm_enabled: true
helm_addons_enabled: true
helm_addons_values:
  ingress-nginx:
    enabled: true
```

To speed-up deployment, consider adding `-t helm -t helm_addons` at the
end of the Ansible command-line.